% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Ein Zufallsgenerator für korrelierte Variablen}
\usage{
correlation(json)
}
\arguments{
\item{json}{ein JSON-Objekt mit den Eigenschaften „len“ (die Stichprobengröße, Standardwert ist 30) und „r“ (gewünschte Korrelation, Standardwert ist 0).}
}
\value{
Ein JSON-Objekt mit den Eigenschaften „table“ und „regression“. Data enthält die Arrays „x“ und „y“, Regression ist ein Objekt mit den Zahlen „x“ und „y“.
}
\description{
Diese Funktion generiert Zufallsvariablen mit gegebener Pearson-Korrelation und Länge, sowie die extrapolierten Extrempunkte für eine Regressionsgerade.
}

